Fix controllers/clients naming
Remove exception handling on client level

On branch MR-31-user-controller
