+++ shared/src/commonMain/kotlin/org/itmodreamteam/myrest/shared/restaurant/RestaurantClient.kt
@@ -3,7 +3,7 @@ package org.itmodreamteam.myrest.shared.restaurant
 import org.itmodreamteam.myrest.shared.ContentPage
 import org.itmodreamteam.myrest.shared.Pageable
 
-interface RestaurantController {
+interface RestaurantClient {
 
     suspend fun search(keyword: String, pageable: Pageable): ContentPage<RestaurantInfo>
 }
